logs: rename to logs to avoid namespace conflict
