+++ logs/logs.go
@@ -1,4 +1,4 @@
-package log
+package logs
 
 import (
 	"fmt"
