+++ logs/logs_test.go
@@ -1,4 +1,4 @@
-package log
+package logs
 
 import (
 	"testing"
